LaTeX_Tab
This function creates a LaTeX formatted table ready to be copied straight into the LaTeX file. The data is input via a list of arrays on same length. The physical quantities and units for the corresponding columns have to be input via a list each. Name has to be a string without suffix. This code can be run to see an example table of random values.